:recycle: Address PR #229 review: relocate forecast Exception + dataclass
- Move ProjectStartDateRequiredError from services/forecast.py to
  projects/exceptions.py (next to ProjectColumnSetError, ProjectDatesError,
  TaskStatusError, GithubMilestoneAlreadyExistsError).
- Move the previously-private _UserContext dataclass to projects/definitions.py
  as ProjectAssignmentForecastUserContext, consistent with the existing
  ProjectProgressStatus dataclass already living there.
- forecast.py now imports both from their new homes; __all__ added so
  external callers (viewset, admin, tests) keep importing
  ProjectStartDateRequiredError from services.forecast unchanged.

Refs monkut/kippo#229 review (kippo#226 / Phase 1 of #224).
