+++ kippo/projects/exceptions.py
@@ -12,3 +12,7 @@ class TaskStatusError(Exception):
 
 class GithubMilestoneAlreadyExistsError(Exception):
     pass
+
+
+class ProjectStartDateRequiredError(ValueError):
+    """Raised when a forecast/suggest operation is requested for a project with no start_date."""
